Remove all uses of unimplemented
